+++ src/Provider/Eng/City.php
@@ -7,8 +7,7 @@ use Fauxify\Fauxify\Provider\ParentCity;
 class City extends ParentCity{
 
     protected $city = [
-        'Kathmandu', 'Pokhara', 'Biratnagar', 'Birgunj', 'Dharan', 'Bharatpur', 'Janakpur', 'Hetauda', 'Butwal', 'Nepalgunj', 'Itahari', 'Dhangadhi', 'Bhaktapur', 'Dipayal', 'Dhankuta', 'Gaur', 'Gorkha', 'Ilam', 'Inaruwa', 'Jaleshwar', 'Lahan', 'Lalitpur', 'Mahendranagar', 'Nepalganj', 'Rajbiraj', 'Ramechhap', 'Sankhuwasabha', 'Siraha', 'Tansen', 'Tulsipur', 'Tikapur', 'Tilottama', 'Tulsipur', 'Waling'
+        'Kathmandu', 'Pokhara', 'Biratnagar', 'Birgunj', 'Dharan', 'Bharatpur', 'Janakpur', 'Hetauda', 'Butwal', 'Nepalgunj', 'Itahari', 'Dhangadhi', 'Bhaktapur', 'Dipayal', 'Dhankuta', 'Gaur', 'Gorkha', 'Ilam', 'Inaruwa', 'Jaleshwar', 'Lahan', 'Lalitpur', 'Mahendranagar', 'Nepalganj', 'Rajbiraj', 'Ramechhap', 'Sankhuwasabha', 'Siraha', 'Tansen', 'Tikapur', 'Tilottama', 'Tulsipur', 'Waling'
     ];
     
-
 }