+++ src/Provider/Eng/Payment.php
@@ -7,7 +7,7 @@ use Fauxify\Fauxify\Provider\ParentPayment;
 class Payment extends ParentPayment{
 
     protected $payment = [
-        'Esewa', 'Khalti', 'IME Pay', 'Connect IP', 'NTC Pay'
+        'Esewa', 'Khalti', 'IME Pay', 'Connect IPS', 'NTC Pay'
     ];
 
 }