+++ src/Provider/Nep/Payment.php
@@ -7,7 +7,7 @@ use Fauxify\Fauxify\Provider\ParentPayment;
 class Payment extends ParentPayment{
 
     protected $payment = [
-        'Esewa Nepali', 'Khalti Nepali', 'IME Pay Nepali', 'Connect IP Nepali', 'NTC Pay Nepali'
+        'इ-सेवा', 'खल्ती', 'आइएमई पे', 'कनेक्ट आइपीएस', 'एनटीसी पे'
     ];
 
 }