+++ src/Provider/ParentPayment.php
@@ -10,28 +10,17 @@ class ParentPayment extends Base{
         'Esewa', 'Khalti', 'IME Pay', 'Connect IP', 'NTC Pay'
     ];
 
-    protected static $methodCallCount = -1;
-    
-    protected $isSuffle = FALSE;
-    protected $shuffled = [];
-
     public function payment($unique = NULL)
     {
 
-        // if($unique == 'UNIQUE')
-        // {
-        //     if($this->isSuffle == FALSE)
-        //     {
-        //         shuffle($this->payment);
-        //         $this->shuffled = $this->payment;
-        //         $this->isSuffle = TRUE;
-        //     }
-        //     self::$methodCallCount++;
-        //     return $this->shuffled[self::$methodCallCount];
-
-        // } else {
-        //     return $this->randArray($this->payment);
-        // }
+        if($unique == 'UNIQUE')
+        {
+            $this->shuffleArray(__FUNCTION__, $this->payment);
+            $this->incrementMethodCallCount(__FUNCTION__);
+            return $this->shuffled[__FUNCTION__][$this->methodCallCount[__FUNCTION__]] ?? NULL;
+        } else {
+            return $this->randArray($this->payment);
+        }
         
         return $this->randArray($this->payment);
     }