+++ test.php
@@ -1,5 +1,0 @@
-<?php
-
-$a = "";
-
-echo isset($a) ? $a : "tala";