+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog for [Fauxify]
 
-## [v1.0.0] - [Release Date]
+## [V1.0.3] - [Feb 1, 2024]
+
+## Added
+- [New] Payment Provider Added. Support both Nepali and English.
+- Now user can generate data uniquely.
+- More data added for person.
+- More optimize.
+
+## Fix
+- If user try to access data beyond the array size using a loop, it now gives null.
+
+
+## [v1.0.0] - [Jan 29, 2024]
 
 ### Added
 - Name: [Generate random name. It support both Nepali and English.]